# OpenAPI.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition for SDC Helm validator
  description: Application for validating Helm charts.
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /validate:
    post:
      tags:
        - ValidationService
      summary: Validate chart
      description: Web endpoint for Helm charts validation.
      operationId: validate
      parameters:
        - name: versionDesired
          in: header
          description: Desired Helm version which should be used to validate the chart.
            If parameter is not provided validation is processing with version based
            on chart's apiVersion. Version could be provided in 'semantic version' or
            'major version'.
          required: false
          schema:
            type: string
          examples:
            Semantic Version:
              description: Semantic version in X.Y.Z format.
              value: 3.5.2
            Major version:
              description: Major version in vX format. Version is mapped to latest available
                major version.
              value: v3
            None:
              description: Request could be sent without versionDesired parameter.
        - name: isLinted
          in: header
          description: "If false, there will be an attempt to render the chart without\
          \ linting it first"
          required: false
          schema:
            type: boolean
            default: false
        - name: isStrictLinted
          in: header
          description: Linting should be strict or not
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: Helm chart that should be validated (packed in .tgz
                    format)
                  format: binary
      responses:
        "400":
          description: Chart cannot be validated using selected version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        "200":
          description: Helm chart successfully validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
              examples:
                Simple Validation:
                  $ref: '#/components/examples/simpleValidation'
                Lint Validation:
                  $ref: '#/components/examples/validationWithLint'
        "500":
          description: Something went wrong during validation execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
components:
  schemas:
...
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
    ValidationResult:
      type: object
      properties:
        renderErrors:
          type: array
          items:
            type: string
        lintWarning:
          type: array
          items:
            type: string
        lintError:
          type: array
          items:
            type: string
        versionUsed:
          type: string
        valid:
          type: boolean
        deployable:
          type: boolean
...
  examples:
    simpleValidation:
      description: Example response when parameter isLinted is set to false
      value: "{\"renderErrors\":[],\"versionUsed\":\"3.5.2\",\"deployable\":true}"
    validationWithLint:
      description: Example response when parameter isLinted is set to true
      value: "{\"renderErrors\":[],\"lintWarning\":[],\"lintError\":[],\"versionUsed\"\
        :\"3.5.2\",\"valid\":true,\"deployable\":true}"